Minor appearence change
Fixed issue with items in masthead being outside the container column,
need to fix meta to the right.
Updated banner image to reduce graining.
Change banner to be fixed center top rather than center center which
gets rid of the strange issue when resizing.
Replaced the green link colour with one more in line with the banner.